plot: move template to repo/plot
