+++ dvc/repo/plot/template.py
